quick_starts/weather_service/pyproject.toml
[project]
name = "weather-service"
version = "0.1.0"
description = "A MCP server project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
 "httpx>=0.28.1",
 "mcp>=1.12.4",
 "python-dotenv>=1.1.1",
]
[[project.authors]]
name = "kr-MATAGI"
email = "dev_matagi@kakao.com"

[build-system]
requires = [ "uv_build>=0.8.6,<0.9.0",]
build-backend = "uv_build"

[project.scripts]
weather-service = "weather_service:main"
